Optimize solver
