+++ src/solver.test.ts
@@ -1,26 +1,56 @@
-import { solve, seqToString } from './solver'
+import { solve } from './solver'
+import { originalSolve } from './test/original-solver'
 
-const matrix = [
-  ['e9', '1c', '1c', '1c', 'bd', 'e9', '55'],
-  ['ff', '1c', 'bd', '55', '1c', 'e9', '7a'],
-  ['1c', '55', '7a', '1c', 'bd', 'bd', 'ff'],
-  ['7a', '7a', '7a', '55', 'ff', '1c', 'bd'],
-  ['1c', '7a', '7a', '55', 'ff', '55', '1c'],
-  ['1c', '1c', '7a', 'bd', '1c', 'bd', '55'],
-  ['bd', 'ff', '7a', '55', 'bd', 'bd', '7a'],
+const cases = [
+  {
+    matrix: [
+      ['e', '1', '1', '1', 'b', 'e', '5'],
+      ['f', '1', 'b', '5', '1', 'e', '7'],
+      ['1', '5', '7', '1', 'b', 'b', 'f'],
+      ['7', '7', '7', '5', 'f', '1', 'b'],
+      ['1', '7', '7', '5', 'f', '5', '1'],
+      ['1', '1', '7', 'b', '1', 'b', '5'],
+      ['b', 'f', '7', '5', 'b', 'b', '7'],
+    ],
+    targets: [
+      ['f', '7', '1'],
+      ['5', '5', '1', 'b'],
+      ['b', '1', '1', 'e'],
+    ],
+  },
+  {
+    matrix: [
+      ['1', '5', 'f', 'b', 'e'],
+      ['b', '1', 'e', 'f', 'e'],
+      ['5', 'b', 'f', '1', '1'],
+      ['e', 'b', '1', '5', '5'],
+      ['5', 'e', 'b', '5', 'f'],
+    ],
+    targets: [
+      ['e', '5'],
+      ['5', 'b', 'e'],
+      ['f', '1', 'b', 'e'],
+      ['5', '1', 'f', '5'],
+    ],
+  },
 ]
 
-const targets = [
-  ['ff', '7a', '1c'],
-  ['55', '55', '1c', 'bd'],
-  ['bd', '1c', '1c', 'e9'],
-]
-
-test('solve 7x7', () => {
-  const results = [4, 5, 6, 7, 8].map(i =>
-    solve(matrix, targets, i).sort(({ seq: seqA }, { seq: seqB }) =>
-      seqToString(seqA).localeCompare(seqToString(seqB))
+test('Solver', () => {
+  const lengths = [4, 5, 6, 7, 8]
+  cases.forEach(({ matrix, targets }) => {
+    const originalResults = lengths.map(i =>
+      originalSolve(matrix, targets, i)
+        .map(res => ({
+          ...res,
+          seq: res.seq.map(([row, col]) => `${row}${col}`).join(''),
+        }))
+        .sort(({ seq: seqA }, { seq: seqB }) => seqA.localeCompare(seqB))
     )
-  )
-  expect(results).toMatchSnapshot()
+    const results = lengths.map(i =>
+      solve(matrix, targets, i).sort(({ seq: seqA }, { seq: seqB }) =>
+        seqA.localeCompare(seqB)
+      )
+    )
+    expect(results).toEqual(originalResults)
+  })
 })
